# orchestrator/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Yandex Lyceum Calculator API",
        "title": "Orchestrator API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/expressions/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expressions"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "01JTE60CDWQ5R3QSWZBP8J6FG3",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetByIdExpressionResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ApiError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/dto.ApiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.ApiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "message": {
                    "type": "string",
                    "example": "error: nil ptr dereference"
                }
            }
        },
...
        "dto.Expression": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "01JTE60CDWQ5R3QSWZBP8J6FG3"
                },
                "result": {
                    "type": "number",
                    "example": 4
                },
                "status": {
                    "type": "string",
                    "example": "DONE"
                }
            }
        },
        "dto.GetByIdExpressionResponse": {
            "type": "object",
            "properties": {
                "expression": {
                    "$ref": "#/definitions/dto.Expression"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "\"Введите 'Bearer TOKEN' чтобы правильно использовать JWT API Token\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}